requirements.txt
torch>=1.10.0
torchvision
Pillow
PyYAML
tqdm
